Skip temporary update (on release_info) if it has already done

+++ release_stuff/scripts/release_info.sh
@@ -6,6 +6,7 @@ RELEASE_BASE_DIR="$HOME_DIR/openmrs-eip-docker"
 RELEASE_DIR="$RELEASE_BASE_DIR/release_stuff"
 RELEASE_SCRIPTS_DIR="$RELEASE_DIR/scripts"
 ONGOING_UPDATE_INFO_FILE="$HOME_DIR/ongoing_update_info"
+TMP_UPDATE_DONE="$HOME_DIR/tmp_update_done_01"
 
 export RELEASE_NAME="EIP release 2.0.0.1"
 export RELEASE_DATE="2022-03-03 9:05:00"
@@ -20,4 +21,8 @@ cd $RELEASE_SCRIPTS_DIR
 
 rm $ONGOING_UPDATE_INFO_FILE
 
-./updates.sh 2>&1 | tee -a $LOG_DIR/upgrade.log
+if [ -f "$TMP_UPDATE_DONE" ]; then 
+	echo "THE TMP UPDATE HAS ALREADY DONE"
+else
+	./updates.sh 2>&1 | tee -a $LOG_DIR/upgrade.log
+fi
